Fixed XR Character after Toolkit Update
- Replaced deprecated "MyMoveController.cs" on "XR Origin" with "PlayerController.cs"
- Replaced Ray Interaction with Direct Interaction on both hands
- Set up "lectureController.cs" correctly for both hands
